Feature suggestion: Add CC license to photo (#18)
* inital UI update for licensing

* Tweaked the `.select` class styles

Made the select box a little wider to show more of the option text.
Also made the margin larger to space elements out a little more
nicely.

* Select element for adding a license

Created a hard-coded select element on the license modal. Users are
prompted to choose a Creative Commons license when the modal
is launched. They are also provided a link out to the Creative
Commons website to help them choose a license if they've never
used one before.

* Add license info to sidebar

Added the license info to the sidebar with `photo.json.license`.

* Made requested changes in #18

- Added `None` as a license option and localized
- Localized help URL text