Fix issue https://github.com/LycheeOrg/Lychee/issues/120 (#19)
* Fix issue https://github.com/LycheeOrg/Lychee/issues/120

The select box is in a modal, so the `select` element can't be targeted
in `view.photo` because it doesn't exist at load.

`basicModal` supports a callback function. Setting the current license
is done from `photo.json.license` after the modal is visible and called
by the callback function.

I've seen an intermittent flash on the screen when setting a license or
cancelling a license change on `basicModal.` I can't identify a cause
in the debugger or watching the call stack. Can you try to reproduce?