Assorted multiselect fixes
Rational support for "Select all" in albums with photos and subalbums
(do not generate an error message).

Add support for Shift key to select a range of albums or photos.

Intuitive clearing of selection with a single click.

multiselect.removeItem() didn't work correctly for non-first items.

Cleaned up calculations of selection and overlap.

New multiselect.unbind() function to match the existing
multiselect.bind().