Make the back button return to search results
Make search results behave a lot more like a smart album.

Set album.json.id for search results and treat it as a legitimate album
ID.

Override parent_id for albums in search results.

Replace / with = in search URL so as not to confuse the search term with
photo ID.