fix dropbox context menu entry (#81) (#255)
* fix dropbox context menu entry

* suppress 'import from dropbox' in context menu if no key is set
* do not throw error if dropbox key is cleared (set to empty)

* prevent 'null' string as value for undefined settings

details: clearing the dropbox key on the settings page (not full_settings) puts null into the config table of the database. on the next reading of the configuration this result in the string 'null' for the `dropbox_key` on the full_settings page (but not on the common settings page).
This fix ensures that null values and empty strings are treated as the same parameter value.