improve support for touch devices (#254)
* refactor :hover CSS
* restrict :hover CSS to devices that support it (this resolves unintuitive behaviour on touch devices)

* reorganize management of image overlays

for touch devices it is inconvenient (and for devices with a mouse unnecessary) to use two separate mechanisms to toggle overlays (mouse clicks) and to cycle through potentially different types of overlays (`o` key).
* add new type of overlay `none` and include it in cycle
* change cycle to `desc`, `date`, `exif`, `none`
* mouse clicks, `o` key, and tap all cycle through overlays
* method renamed from `photo.update_display_overlay` to `photo.cycle_display_overlay`
* `imageview.on("click", "img", photo.cycle_display_overlay)` in `$(document).ready(...)` (`view/main.js`) not necessary (already bound in `init.js`); avoids code duplication.
* `image_overlay` config no longer needed. set 'image_overlay_type' to 'none' or the desired type instead (addressed in separate commit)

* retire setting `image_overlay`
* always show title in overlay (if visible)

`mousemove` or touch removed title of overlay (if visible) but not the camera info etc and re-displayed the title after some delay (when hiding menu bar). this was somewhat confusing - especially on touch devices. overlay now remains unchanged.

* increase size of back button

on small touch devices the back button is easily missed and instead the photo/album title is activated.

* size of back button for albums and settings

compensate for the removed padding of the menu bar on the left by increasing size of back/settings buttons

* increase context menu entries on touch devices
* add some visual feedback when hovering over context menu
* minor layout improvements to settings pages
* mark buttons in color on touch devices
* add space to buttons and inputs
* simplify _settings.scss logic

* tweaks to buttons in header

* remove left/right padding of header and increase size of settings/back/add buttons