Add album tracks
Adds a new feature that allows to upload GPX tracks/files and link them to an album (only regular, no smart/tag albums). This track is displayed together with the photos on the map, means this is only usable if the map is enabled.
To display the track, the Leaflet plugin https://github.com/mpetazzoni/leaflet-gpx is used, it is very easy and only needs one file.